Revert "Temporarily Revert "Reorganize code base""
