+++ src/main/java/seedu/address/logic/parser/TagCommandParser.java
@@ -10,7 +10,7 @@ import java.util.Set;
 
 import javafx.util.Pair;
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.TagCommand;
+import seedu.address.logic.commands.tag.TagCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.tag.Tag;
