+++ src/test/java/seedu/address/logic/commands/CommandTestUtil.java
@@ -17,6 +17,7 @@ import java.util.List;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.commands.person.EditCommand;
 import seedu.address.logic.commands.wedding.EditWeddingCommand;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
