+++ src/test/java/seedu/address/logic/parser/tag/CreateTagCommandParserTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.tag;
 
 import static seedu.address.logic.commands.CommandTestUtil.INVALID_TAG_DESC;
 import static seedu.address.logic.commands.CommandTestUtil.PREAMBLE_WHITESPACE;
@@ -11,7 +11,8 @@ import static seedu.address.logic.parser.CommandParserTestUtil.assertParseSucces
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.Messages;
-import seedu.address.logic.commands.CreateTagCommand;
+import seedu.address.logic.commands.tag.CreateTagCommand;
+import seedu.address.logic.parser.CreateTagCommandParser;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.tag.TagName;
